Safer version of deleting the existing addon dir

+++ WowUp.WPF/Services/AddonService.cs
@@ -19,6 +19,7 @@ using WowUp.WPF.Services.Contracts;
 using WowUp.WPF.Utilities;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Processing;
+using Serilog;
 
 namespace WowUp.WPF.Services
 {
@@ -413,12 +414,44 @@ namespace WowUp.WPF.Services
                 {
                     var unzippedDirectoryName = Path.GetFileName(unzippedFolder);
                     var unzipLocation = Path.Combine(addonFolderPath, unzippedDirectoryName);
+                    var unzipBackupLocation = Path.Combine(addonFolderPath, $"{unzippedDirectoryName}-bak");
+
+                    // If the user already has the addon installed, create a temporary backup
                     if (Directory.Exists(unzipLocation))
                     {
-                        await FileUtilities.DeleteDirectory(unzipLocation);
+                        Directory.Move(unzipLocation, unzipBackupLocation);
                     }
 
-                    FileUtilities.DirectoryCopy(unzippedFolder, unzipLocation);
+                    try
+                    {
+                        // Copy contents from unzipped new directory to existing addon folder location
+                        FileUtilities.DirectoryCopy(unzippedFolder, unzipLocation);
+
+                        // If the copy succeeds, delete the backup
+                        if (Directory.Exists(unzipBackupLocation))
+                        {
+                            await FileUtilities.DeleteDirectory(unzipBackupLocation);
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Log.Error(ex, $"Failed to copy addon directory {unzipLocation}");
+                        // If a backup directory exists, attempt to roll back
+                        if (Directory.Exists(unzipBackupLocation))
+                        {
+                            // If the new addon folder was already created delete it
+                            if (Directory.Exists(unzipLocation))
+                            {
+                                await FileUtilities.DeleteDirectory(unzipLocation);
+                            }
+
+                            // Move the backup folder into the original location
+                            Log.Information($"Attempting to roll back {unzipBackupLocation}");
+                            Directory.Move(unzipBackupLocation, unzipLocation);
+                        }
+
+                        throw;
+                    }
                 }
             });
         }
